adding address CRUD and jwtAuth system

+++ src/main/java/br/univesp/ocorrencia_api/service/ClientService.java
@@ -3,8 +3,8 @@ package br.univesp.ocorrencia_api.service;
 import br.univesp.ocorrencia_api.entity.Client;
 import br.univesp.ocorrencia_api.exception.DuplicatedCpfException;
 import br.univesp.ocorrencia_api.repository.ClientRepository;
-import br.univesp.ocorrencia_api.usecases.clientusecases.ClientResponseDTO;
-import br.univesp.ocorrencia_api.usecases.clientusecases.CreateClientDTO;
+import br.univesp.ocorrencia_api.usecases.clientusecases.ClientResponse;
+import br.univesp.ocorrencia_api.usecases.clientusecases.ClientRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -20,36 +20,45 @@ public class ClientService {
         this.clientRepository = clientRepository;
     }
 
-    public ClientResponseDTO create(CreateClientDTO createClientDTO) {
-        if (clientRepository.existsByCpf(createClientDTO.cpf())) {
+    public ClientResponse create(ClientRequest clientRequest) {
+        if (clientRepository.existsByCpf(clientRequest.cpf())) {
             throw new DuplicatedCpfException("Client already exists");
         }
-        Client client = Client.fromRequest(createClientDTO.name(), createClientDTO.birthDate(), createClientDTO.cpf());
+        Client client = Client.fromRequest(clientRequest.name(), clientRequest.birthDate(), clientRequest.cpf());
         clientRepository.save(client);
-        return ClientResponseDTO.fromEntity(client);
+        return ClientResponse.fromEntity(client);
     }
 
-    public Page<ClientResponseDTO> findAll(Pageable pageable) {
-        return clientRepository.findAll(pageable).map(ClientResponseDTO::fromEntity);
+    public Page<ClientResponse> findAll(Pageable pageable) {
+        Page<Client> clients = clientRepository.findAll(pageable);
+        if (clients.isEmpty()) {
+            return Page.empty(pageable);
+        }
+        return clients.map(ClientResponse::fromEntity);
     }
 
-    public Optional<ClientResponseDTO> findById(Long id) {
-        return clientRepository.findById(id).map(ClientResponseDTO::fromEntity);
+    public Optional<ClientResponse> findById(Long id) {
+        return clientRepository.findById(id).map(ClientResponse::fromEntity);
     }
 
-    public ClientResponseDTO update(Long id, CreateClientDTO createClientDTO) {
+    public ClientResponse update(Long id, ClientRequest clientRequest) {
         Client client = clientRepository.findById(id).orElseThrow();
-        if (clientRepository.existsByCpf(createClientDTO.cpf())) {
+        if (clientRepository.existsByCpf(clientRequest.cpf())) {
             throw new DuplicatedCpfException("Client already exists");
         }
-        client.setName(createClientDTO.name());
-        client.setBirthDate(createClientDTO.birthDate());
-        client.setCpf(createClientDTO.cpf());
+        client.setName(clientRequest.name());
+        client.setBirthDate(clientRequest.birthDate());
+        client.setCpf(clientRequest.cpf());
         clientRepository.save(client);
-        return ClientResponseDTO.fromEntity(client);
+        return ClientResponse.fromEntity(client);
     }
 
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (clientRepository.existsById(id)) {
+            clientRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Client not found");
+        }
+
     }
 }
